requirements.txt
requests==2.31.0
beautifulsoup4==4.12.0
fpdf==1.7.2
discord.py